requirements.txt
apache-airflow==2.7.2
clickhouse-connect==0.6.14
clickhouse-driver==0.2.6
pandas==2.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.22